Split server internal modules
